+++ backend/app/models/auth_session.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 
-from sqlalchemy import DateTime, Integer, String, func
+from sqlalchemy import BigInteger, DateTime, Integer, String, func
 from sqlalchemy.orm import Mapped, mapped_column
 
 from app.db.base import Base
@@ -12,7 +12,7 @@ class WebAuthSession(Base):
     id: Mapped[str] = mapped_column(String(64), primary_key=True)
     csrf_token: Mapped[str | None] = mapped_column(String(128), nullable=True)
     link_code: Mapped[str | None] = mapped_column(String(16), unique=True, nullable=True, index=True)
-    telegram_user_id: Mapped[int | None] = mapped_column(Integer, nullable=True, index=True)
+    telegram_user_id: Mapped[int | None] = mapped_column(BigInteger, nullable=True, index=True)
     telegram_username: Mapped[str | None] = mapped_column(String(80), nullable=True)
     telegram_display_name: Mapped[str | None] = mapped_column(String(120), nullable=True)
     link_code_expires_at: Mapped[datetime | None] = mapped_column(DateTime(timezone=True), nullable=True)
